Validação pelo nome

+++ projeto-dio-padroes/src/main/java/com/dio/project/padroes/java/service/ClienteServiceImpl.java
@@ -1,11 +1,11 @@
 package com.dio.project.padroes.java.service;
 
 import com.dio.project.padroes.java.error.NotFoundException;
+import com.dio.project.padroes.java.error.ClienteAlreadyExistsException;
 import com.dio.project.padroes.java.model.Cliente;
 import com.dio.project.padroes.java.model.Endereco;
 import com.dio.project.padroes.java.repository.ClienteRepository;
 import com.dio.project.padroes.java.repository.EnderecoRepository;
-import org.aspectj.weaver.ast.Not;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -29,9 +29,11 @@ public class ClienteServiceImpl implements ClienteService {
     @Override
     public Cliente buscarPorId(Long id) {
         Optional<Cliente> cliente = clienteRepository.findById(id);
+
         if (cliente.isPresent()) {
             return cliente.get();
-        } throw new NotFoundException("Cliente não encontrado!");
+        }
+        throw new NotFoundException("Cliente não encontrado!");
     }
 
     @Override
@@ -43,21 +45,46 @@ public class ClienteServiceImpl implements ClienteService {
         Optional<Cliente> clienteBd = clienteRepository.findById(id);
         if (clienteBd.isPresent()) {
             salvarClienteComCep(cliente);
+        } else {
+            throw new NotFoundException("Cliente não encontrado!");
         }
+
     }
     @Override
     public void deletar(Long id) {
+        Optional<Cliente> cliente = clienteRepository.findById(id);
+        if(cliente.isPresent()) {
+            clienteRepository.deleteById(id);
+        } else {
+            throw new NotFoundException("Cliente não encontrado!");
+        }
 
-        clienteRepository.deleteById(id);
     }
     private void salvarClienteComCep(Cliente cliente) {
-        String cep = cliente.getEndereco().getCep();
-        Endereco endereco = enderecoRepository.findById(cep).orElseGet(() -> {
-            Endereco novoEndereco = viaCepService.consultarCep(cep);
-            enderecoRepository.save(novoEndereco);
-            return novoEndereco;
-        });
-        cliente.setEndereco(endereco);
-        clienteRepository.save(cliente);
+        Boolean clienteDb = validarPeloNome(cliente.getNome());
+        if(!clienteDb){
+            String cep = cliente.getEndereco().getCep();
+            Endereco endereco = enderecoRepository.findById(cep).orElseGet(() -> {
+                Endereco novoEndereco = viaCepService.consultarCep(cep);
+                enderecoRepository.save(novoEndereco);
+                return novoEndereco;
+            });
+            cliente.setEndereco(endereco);
+            clienteRepository.save(cliente);
+        } else {
+            throw new ClienteAlreadyExistsException("O cliente %s já existe no Banco de Dados".formatted(cliente.getNome()));
+        }
+
+
+    }
+
+    private Boolean validarPeloNome(String nome) {
+        Boolean validacao = false;
+        for(Cliente cliente: clienteRepository.findAll()) {
+            if(cliente.getNome().equals(nome)) {
+                validacao = true;
+            }
+        }
+        return validacao;
     }
 }
